REF/TST: misplaced MultiIndex tests (#32314)
